Move basic http auth and cookies in submodules
